+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `concat` is now variadic and accepts one or more string arguments
+  (`concat("a")` returns `"a"`)
+
 ## [0.2.2] - 2026-08-02
 
 ### Added